Remove redundant fromIntegral

+++ Set2/Challenge10.hs
@@ -13,6 +13,6 @@ main = do
     file <- readFile "Set2/data/10.txt"
     let decodedFile = B.unpack $ C.pack $ decodeBase64 $ filter (/= '\n') file
     let key = B.unpack $ C.pack "YELLOW SUBMARINE"
-    let iv = replicate 16 (fromIntegral 0x00)
+    let iv = replicate 16 0x00
     let decrypted = decrypt_AES128_CBC iv key decodedFile
     putStrLn $ C.unpack $ B.pack decrypted
